feature/refactor-easy-request-id (#1229)
